+++ common/jwt-common/src/main/java/pro/coachcore/jwt/JwtAuthConverter.java
@@ -4,7 +4,7 @@ import java.util.stream.Collectors;
 
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.lang.Nullable;
-import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.authentication.AuthenticationCredentialsNotFoundException;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
@@ -17,18 +17,16 @@ public class JwtAuthConverter implements Converter<Jwt, JwtAuthenticationToken> 
   @Nullable
   public JwtAuthenticationToken convert(@Nullable Jwt jwt) {
     if (jwt == null) {
-      return null;
+      throw new AuthenticationCredentialsNotFoundException("Jwt token not found");
     }
 
     var claim = jwt.getClaimAsStringList("roles");
 
     if (claim == null) {
-      return null;
+      return new JwtAuthenticationToken(jwt);
     }
 
-    var roles = claim.stream()
-        .map(SimpleGrantedAuthority::new)
-        .collect(Collectors.toList());
+    var roles = claim.stream().map(SimpleGrantedAuthority::new).collect(Collectors.toList());
 
     return new JwtAuthenticationToken(jwt, roles);
   }