+++ training-service/src/main/java/pro/coachcore/training/security/SecurityConfig.java
@@ -1,25 +1,31 @@
 package pro.coachcore.training.security;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.core.convert.converter.Converter;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.oauth2.jwt.Jwt;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.security.oauth2.server.resource.web.authentication.BearerTokenAuthenticationFilter;
 import org.springframework.security.web.SecurityFilterChain;
-import static org.springframework.security.config.Customizer.withDefaults;
+
+import pro.coachcore.jwt.JwtAuthConverter;
 
 @Configuration
 @EnableWebSecurity
-@RequiredArgsConstructor
 public class SecurityConfig {
-  private final TrainingPlanAuthorizationManager trainingPlanAuthorizationManager;
+  @Bean
+  Converter<Jwt, JwtAuthenticationToken> jwtConverter() {
+    return new JwtAuthConverter();
+  }
 
   @Bean
-  SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+  SecurityFilterChain filterChain(HttpSecurity http,
+      TrainingPlanAuthorizationManager trainingPlanAuthorizationManager) throws Exception {
     http.csrf(AbstractHttpConfigurer::disable);
 
     http.authorizeHttpRequests(auth -> {
@@ -29,14 +35,18 @@ public class SecurityConfig {
       auth.requestMatchers("/v1/training-plans/me").authenticated();
       auth.requestMatchers("/v1/training-plans/{planId}/**")
           .access(trainingPlanAuthorizationManager);
-      auth.requestMatchers("/v1/catalog-exercises").permitAll();
+      auth.requestMatchers(HttpMethod.POST, "/v1/catalog-exercises").hasAuthority("ADMIN");
+      auth.requestMatchers(HttpMethod.GET, "/v1/catalog-exercises").permitAll();
 
       auth.requestMatchers(HttpMethod.GET, "/v1/exercise-categories").permitAll();
       auth.requestMatchers(HttpMethod.GET, "/v1/training-plans").permitAll();
       auth.requestMatchers(HttpMethod.GET, "/swagger-ui/**", "/v3/api-docs/**").permitAll();
+      auth.requestMatchers("/test/**").permitAll();
     });
 
-    http.oauth2ResourceServer(server -> server.jwt(withDefaults()));
+    http.oauth2ResourceServer(server -> server.jwt(jwt -> {
+      jwt.jwtAuthenticationConverter(jwtConverter());
+    }));
 
     http.sessionManagement(session -> {
       session.sessionCreationPolicy(SessionCreationPolicy.STATELESS);