+++ training-service/src/test/java/pro/coachcore/training/catalog/category/ExerciseCategoryControllerTests.java
@@ -1,43 +1,27 @@
 package pro.coachcore.training.catalog.category;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.transaction.annotation.Transactional;
 
-import pro.coachcore.common.test.security.jwt.JwtTestContext;
-import pro.coachcore.training.jwt.JwtTestConfig;
-
-import static org.hamcrest.core.Is.is;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
 @Transactional
-@ExtendWith(SpringExtension.class)
 @SpringBootTest
 @AutoConfigureMockMvc
 @AutoConfigureTestDatabase
-@Import(JwtTestConfig.class)
 class ExerciseCategoryControllerTests {
   @Autowired
   private MockMvc mockMvc;
 
-  @Autowired
-  private ObjectMapper objectMapper;
-
-  @Autowired
-  private JwtTestContext jwtTestContext;
-  
   @Test
-  void shouldGetExerciseCategories() throws Exception {
+  void getCategories_shouldGetCategories_forAnyUser() throws Exception {
     mockMvc.perform(get("/v1/exercise-categories"))
         .andExpect(status().isOk())
         .andExpect(jsonPath("$").isArray());