+++ training-service/src/test/java/pro/coachcore/training/catalog/exercise/CatalogExerciseControllerTests.java
@@ -1,34 +1,32 @@
 package pro.coachcore.training.catalog.exercise;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.transaction.annotation.Transactional;
-
-import pro.coachcore.common.test.security.jwt.JwtTestContext;
-import pro.coachcore.training.TestDtos;
-import pro.coachcore.training.catalog.category.ExerciseCategoryDto;
-import pro.coachcore.training.jwt.JwtTestConfig;
-
+import static org.springframework.http.MediaType.APPLICATION_JSON;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
+import org.springframework.security.test.context.support.WithMockUser;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+import pro.coachcore.training.TestConfig;
+import pro.coachcore.training.TestDtoFactory;
+import pro.coachcore.training.catalog.category.ExerciseCategoryDto;
+
 @Transactional
-@ExtendWith(SpringExtension.class)
-@SpringBootTest
-@AutoConfigureMockMvc
+@SpringBootTest(classes = TestConfig.class)
+@Import(TestConfig.class)
 @AutoConfigureTestDatabase
-@Import(JwtTestConfig.class)
+@AutoConfigureMockMvc(addFilters = true)
 class CatalogExerciseControllerTests {
   @Autowired
   private MockMvc mockMvc;
@@ -36,27 +34,37 @@ class CatalogExerciseControllerTests {
   @Autowired
   private ObjectMapper objectMapper;
 
-  @Autowired
-  private JwtTestContext jwtTestContext;
-
   @Test
-  void shouldGetCatalogExercises() throws Exception {
+  void getCatalogExercise_shouldReturn200_forAnyUser() throws Exception {
     mockMvc.perform(get("/v1/catalog-exercises"))
         .andExpect(status().isOk())
         .andExpect(jsonPath("$").isArray());
   }
 
   @Test
-  void shouldPostCatalogExercise() throws Exception {
-    var result = mockMvc.perform(get("/v1/exercise-categories"))
-        .andReturn();
+  void postCatalogExercise_shoulGetCategoryAndReturn401_whenNoAdminRole() throws Exception {
+    var result = mockMvc.perform(get("/v1/exercise-categories")).andReturn();
     var json = result.getResponse().getContentAsString();
     var categoryId = objectMapper.readValue(json, ExerciseCategoryDto[].class)[0].id();
+    var requestBody = TestDtoFactory.createCatalogExerciseDto();
 
     mockMvc.perform(post("/v1/catalog-exercises")
-            .param("categoryId", categoryId)
-            .contentType(MediaType.APPLICATION_JSON)
-            .content(objectMapper.writeValueAsString(TestDtos.CREATE_CATALOG_EXERCISE_DTO)))
-        .andExpect(status().isCreated());
+        .param("categoryId", categoryId)
+        .contentType(APPLICATION_JSON).content(objectMapper.writeValueAsString(requestBody)))
+        .andExpect(status().is(401));
+  }
+
+  @Test
+  @WithMockUser(username = "user", authorities = "ADMIN")
+  void postCatalogExercise_shoulGetCategoryAndReturn201_whenAdminAccess() throws Exception {
+    var result = mockMvc.perform(get("/v1/exercise-categories")).andReturn();
+    var json = result.getResponse().getContentAsString();
+    var categoryId = objectMapper.readValue(json, ExerciseCategoryDto[].class)[0].id();
+    var requestBody = TestDtoFactory.createCatalogExerciseDto();
+
+    mockMvc.perform(post("/v1/catalog-exercises")
+        .param("categoryId", categoryId)
+        .contentType(APPLICATION_JSON).content(objectMapper.writeValueAsString(requestBody)))
+        .andExpect(status().is(201));
   }
 }