+++ training-service/src/test/java/pro/coachcore/training/plan/unit/TrainingUnitServiceTests.java
@@ -1,16 +1,21 @@
 package pro.coachcore.training.plan.unit;
 
 import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.when;
+
 import java.time.DayOfWeek;
-import java.util.UUID;
+
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+
 import pro.coachcore.exception.ResourceAlreadyExistsException;
 import pro.coachcore.exception.ResourceNotFoundException;
+import pro.coachcore.lang.message.MessageService;
+import pro.coachcore.training.TestDtoFactory;
 import pro.coachcore.training.plan.TrainingPlan;
 import pro.coachcore.training.plan.TrainingPlanRepository;
 
@@ -22,12 +27,15 @@ class TrainingUnitServiceTests {
   @Mock
   private TrainingUnitRepository trainingUnitRepository;
 
+  @Mock
+  private MessageService messageService;
+
   @InjectMocks
   private TrainingUnitService trainingUnitService;
 
   @Test
-  void throwsResourceNotFound_findAllTrainingUnitsByTrainingPlanLocalId() {
-    var planId = UUID.randomUUID().toString();
+  void getAllUnits_shouldThrowException_whenParentPlanNotFound() {
+    var planId = "plan";
 
     when(trainingPlanRepository.existsByLocalId(planId)).thenReturn(false);
 
@@ -35,24 +43,27 @@ class TrainingUnitServiceTests {
   }
 
   @Test
-  void throwsResourceAlreadyExists_saveTrainingUnit() {
-    var dto = new CreateTrainingUnitDto("MONDAY", "some notes", "name");
-    var plan = new TrainingPlan();
+  void saveUnit_shouldThrowException_whenAlreadyExists() {
+    var plan = spy(new TrainingPlan());
+    var dto = TestDtoFactory.createUnitDto();
 
-    when(trainingUnitRepository.existsByTrainingPlanAndDayOfWeek(plan, DayOfWeek.MONDAY))
-        .thenReturn(true);
+    when(plan.getLocalId()).thenReturn("plan");
+    when(plan.getWeeks()).thenReturn(2);
+    when(trainingUnitRepository.countByTrainingPlan_LocalIdAndDayOfWeek(plan.getLocalId(),
+        DayOfWeek.valueOf(dto.dayOfWeek())))
+        .thenReturn(2L);
 
     assertThrows(ResourceAlreadyExistsException.class,
         () -> trainingUnitService.saveUnit(plan, dto));
   }
 
   @Test
-  void throwsResourceAlreadyExists_updateTrainingUnit() {
+  void updateUnit_shouldThrowException_whenUnitNotFound() {
     var unit = new TrainingUnit();
-    var dto = new UpdateTrainingUnitDto("MONDAY", "Notes", "Name");
-    var planId = UUID.randomUUID().toString();
+    var dto = TestDtoFactory.updateUnitDto();
+    var planId = "plan";
 
-    when(trainingUnitRepository.existsByTrainingPlan_LocalIdAndDayOfWeek(planId, DayOfWeek.MONDAY))
+    when(trainingUnitRepository.existsByTrainingPlan_LocalIdAndDayOfWeek(planId, DayOfWeek.valueOf(dto.dayOfWeek())))
         .thenReturn(true);
 
     assertThrows(ResourceAlreadyExistsException.class,
@@ -60,14 +71,12 @@ class TrainingUnitServiceTests {
   }
 
   @Test
-  void throwsResourceNotFound_deleteTrainingUnitByTrainingPlanLocalIdAndLocalId() {
-    var planId = UUID.randomUUID().toString();
-    var unitId = UUID.randomUUID().toString();
+  void deleteUnit_shouldThrowException_whenUnitNotFound() {
+    var planId = "plan";
+    var unitId = "unit";
 
-    when(trainingUnitRepository.existsByTrainingPlan_LocalIdAndLocalId(planId, unitId))
-        .thenReturn(false);
+    when(trainingUnitRepository.existsByTrainingPlan_LocalIdAndLocalId(planId, unitId)).thenReturn(false);
 
-    assertThrows(ResourceNotFoundException.class,
-        () -> trainingUnitService.deleteUnit(planId, unitId));
+    assertThrows(ResourceNotFoundException.class, () -> trainingUnitService.deleteUnit(planId, unitId));
   }
 }