+++ CHANGELOG.md
@@ -1,1 +1,14 @@
 # Changelog
+
+`release-1.1.0`
+
+- Fix deleting plans
+- Let only admins edit catalog
+- Fix and add tests for training service
+
+# Todo
+
+`25.03.2025`
+
+- Fix foreign keys when deleting plan `DONE`
+- Update postman endpoints `DONE`